feat: guard master release only

+++ server/server-extend.js
@@ -1,22 +1,3 @@
-require('@babel/polyfill');
-require('@babel/register')();
-
-require('css-modules-require-hook')({
-  extensions: ['.scss'],
-  preprocessCss: (data, filename) =>
-    require('node-sass').renderSync({
-      data,
-      file: filename,
-    }).css,
-  camelCase: true,
-  generateScopedName: '[name]__[local]__[hash:base64:8]',
-});
-
-require('asset-require-hook')({
-  extensions: ['jpg', 'png', 'gif', 'webp'],
-  limit: 8000,
-});
-
 const isSever = require('./is-server');
 const cmd = require('node-cmd');
 const createHandler = require('github-webhook-handler');
@@ -28,10 +9,18 @@ handler.on('error', function(err) {
   process.stderr.write(err);
 });
 
-handler.on('push', function() {
-  process.stdout.write('git push event detected.\n');
-  cmd.run('git clean -f');
-  cmd.run('git pull');
+handler.on('push', function(event) {
+  if (event.payload) {
+    const { ref, commits } = event.payload;
+
+    if (ref === 'refs/heads/master') {
+      if (commits.message === 'build: release') {
+        process.stdout.write('git build-release event detected.\n');
+        cmd.run('git clean -f');
+        cmd.run('git pull');
+      }
+    }
+  }
 });
 
 module.exports = function(app) {
